+++ home/dot-bash/overlays/bashrc-os-linux-ubuntu.sh
@@ -2,9 +2,12 @@
 
 # shellcheck disable=SC1091
 source "$HOME/dot-bash/bashrc-lib.sh" # common functions
-source_script "${HOME}/dot-bash/overlays/bashrc-os-linux.sh"
+
+# shellcheck disable=SC1090
+[[ -f "${HOME}/dot-bash/overlays/bashrc-os-linux.sh" ]] && source "${HOME}/dot-bash/overlays/bashrc-os-linux.sh"
 
 # Load WSL2 config conditionally
 if is_wsl2; then
-  source_script "${HOME}/dot-bash/overlays/bashrc-wsl2.sh"
+  # shellcheck disable=SC1090
+  [[ -f "${HOME}/dot-bash/overlays/bashrc-wsl2.sh" ]] && source "${HOME}/dot-bash/overlays/bashrc-wsl2.sh"
 fi